Code cleaning

+++ include/Eeprom.h
@@ -21,12 +21,13 @@
 
 #ifndef INCLUDE_EEPROM_H_
 #define INCLUDE_EEPROM_H_
+#include <arduino.h>
 
 namespace ESParaSite {
 namespace RtcEeprom {
 void initRtcEeprom();
 void dumpEEPROM(uint16_t, uint16_t);
-int doEepromFirstRead();
+int8_t doEepromFirstRead();
 uint8_t doEepromRead(uint16_t);
 uint8_t doEepromWrite();
 void doEepromFormat(uint8_t);
